chore: adjust directory structure
